requirements.txt
Django==3.2.17
django-allauth==0.54.0
djangorestframework==3.14.0
django-cors-headers==3.14.0
django-silk==5.0.3
drf-yasg==1.21.5
django-filter==22.1
django-currentuser==0.5.3
mysqlclient==2.1.1
redis==4.4.2
django-redis==5.2.0
celery==5.2.7
django_celery_beat==2.4.0
django_celery_results==2.4.0
dj-database-url==1.2.0
sentry-sdk==1.27.1
gunicorn==20.1.0
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.5
python-dotenv==0.21.1
django-debug-toolbar==3.8.1
sql-metadata==2.6.0
kafka-python==2.0.2
djangorestframework-api-key==2.0.0
python-decouple==3.7
djangorestframework-jwt==1.11.0
django-filter==22.1
python-decouple==3.7
django-silk==5.0.3
psycopg2-binary==2.9.5
gunicorn==20.1.0
whitenoise~=6.3.0
